fix: Changed type of id field in sql from INT to INT UNSIGNED
fix: Changed type of join_time field in sql from TIMESTAMP to BIGINT UNSIGNED

+++ Client/src/YiZi/Client/Core/User.cpp
@@ -5,10 +5,10 @@ namespace YiZi::Client
 {
     User* User::s_User = nullptr;
 
-    User::User(const uint32_t id, CStringA phone, CString nickname, const uint32_t join_time, const bool is_admin)
+    User::User(const uint32_t id, CStringA phone, CString nickname, const uint64_t join_time, const bool is_admin)
         : m_Id{id}, m_Phone{std::move(phone)}, m_Nickname{std::move(nickname)}, m_JoinTime{join_time}, m_IsAdmin{is_admin} {}
 
-    User* User::New(const uint32_t id, CStringA phone, CString nickname, const uint32_t join_time, const bool is_admin)
+    User* User::New(const uint32_t id, CStringA phone, CString nickname, const uint64_t join_time, const bool is_admin)
     {
         Delete();
         s_User = new User{id, std::move(phone), std::move(nickname), join_time, is_admin};
