+++ Client/src/YiZi/Client/Core/User.h
@@ -5,17 +5,17 @@ namespace YiZi::Client
     class User
     {
     private:
-        User(uint32_t id, CStringA phone, CString nickname, uint32_t join_time, bool is_admin);
+        User(uint32_t id, CStringA phone, CString nickname, uint64_t join_time, bool is_admin);
 
     public:
         [[nodiscard]] uint32_t GetId() const { return m_Id; }
         [[nodiscard]] const CStringA& GetPhone() const { return m_Phone; }
         [[nodiscard]] const CString& GetNickname() const { return m_Nickname; }
-        [[nodiscard]] uint32_t GetJoinTime() const { return m_JoinTime; }
+        [[nodiscard]] uint64_t GetJoinTime() const { return m_JoinTime; }
         [[nodiscard]] bool GetIsAdmin() const { return m_IsAdmin; }
 
         static bool IsValid() { return s_User != nullptr; }
-        static User* New(uint32_t id, CStringA phone, CString nickname, uint32_t join_time, bool is_admin);
+        static User* New(uint32_t id, CStringA phone, CString nickname, uint64_t join_time, bool is_admin);
         [[nodiscard]] static User* Get() { return s_User; }
         static void Delete()
         {
@@ -27,7 +27,7 @@ namespace YiZi::Client
         const uint32_t m_Id;
         const CStringA m_Phone;
         const CString m_Nickname;
-        const uint32_t m_JoinTime;
+        const uint64_t m_JoinTime;
         const bool m_IsAdmin;
 
         static User* s_User;
