+++ Server/src/YiZi/Server/Utils/ClientInfo.h
@@ -11,7 +11,7 @@ namespace YiZi::Server
         const uint32_t id;
         const std::string phone;
         const std::u16string nickname;
-        const uint32_t join_time;
+        const uint64_t join_time;
         const bool is_admin;
 
         SAcceptSocket* const client;
